Add start and end print macros
Started with the sample Macros config, selectively merged in parts from
both Cura and OrcaSlicer's Ender 3 V2 profiles, then heavily adapted the
result to work in Klipper and order the steps slightly differently.

OrcaSlicer's END_PRINT script made use of the `min` function that isn't
available for use in Klipper. It's not as pretty, but it seems to work.